add : app using flask
it's a simple app created using a flask in which we can search for student id and course id of data and if we enter the student details it will check if the student is present it will give the marks of the student and his total. if will enter course id it will give the average and max marks present in the course. if we enter invalid inputs it will throw an error.